feat(gaussianpicker): Review thresholds
Review how the signal and noise thresholds are determined for picking.

Noise
The noise threshold is determined from data outwith the identified pick
windows. Previously, it was simply some percentile of the data, tuned
with the 'picking_threshold' parameter (a value between 0 and 1), that
fixed some percentile of the noise data to be the threshold. This was
severely affected by phase arrivals unrelated to the event in question
creating large peaks in the onsets in the noise window. This wasn't
simply remedied by choosing a lower percentile, as when there was just
noise in the noise window, you ended up with a low threshold and any
peak would be picked. This could be avoided by choosing a percentile
greater than the 88th, as this was used when finding the signal
threshold.

Instead, we turn again to the Median Absolute Deviation - one parameter
is exchanged for another, but this is at least more obviously tunable.
However, it remains to be tested on a wider range of datasets.

Signal
The signal threshold was set simply as the 88th percentile of the data
within the signal window. This would often not cope well with S phase
onsets, for which the peaks are commonly wider than the P phase onset
counterpart. A wider peak means the threshold goes up, and could
sometimes mean an artificially narrow fit was made, giving unrealistic
uncertainties. To remedy, we cap the signal threshold to be no greater
than the half-width of the onset peak in the signal window. If there is
just noise, the noise threshold takes over. If there is signal, we find
a generally better fit is made, resulting in better picks.

This will need to be reviewed carefully across a range of scenarios,
preferably with different noise characteristics - high noise, low noise,
lots of closely spaced events, etc.